move getCpuProfileTimeline from VmServiceWrapper to CpuProfileModel
